+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [1.1.0] - 2024-08-28
+
+Replace ping with heartbeat in the code and .ini
+
+### Changed
+
+- Code, config.ini and test_child.py have been refactored
+
+## [1.0.0] - 2024-06-07
 
 - Initial release